+++ lib/auto_stop_handler.dart
@@ -1,5 +1,6 @@
-import 'package:background_locator/background_locator.dart';
 import 'package:flutter/material.dart';
+
+import 'background_locator.dart';
 
 class AutoStopHandler extends WidgetsBindingObserver {
   @override