+++ lib/callback_dispatcher.dart
@@ -1,10 +1,10 @@
 import 'dart:ui';
 
-import 'package:background_locator/location_dto.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter/widgets.dart';
 
 import 'keys.dart';
+import 'location_dto.dart';
 
 @pragma('vm:entry-point')
 void callbackDispatcher() {
@@ -26,6 +26,17 @@ void callbackDispatcher() {
           PluginUtilities.getCallbackFromHandle(CallbackHandle.fromRawHandle(
               args[Keys.ARG_NOTIFICATION_CALLBACK]));
       notificationCallback();
+    } else if (Keys.BCM_INIT == call.method) {
+      final Map<dynamic, dynamic> args = call.arguments;
+      final Function initCallback = PluginUtilities.getCallbackFromHandle(
+          CallbackHandle.fromRawHandle(args[Keys.ARG_INIT_CALLBACK]));
+      Map<dynamic, dynamic> data = args[Keys.ARG_INIT_DATA_CALLBACK];
+      initCallback(data);
+    } else if (Keys.BCM_DISPOSE == call.method) {
+      final Map<dynamic, dynamic> args = call.arguments;
+      final Function disposeCallback = PluginUtilities.getCallbackFromHandle(
+          CallbackHandle.fromRawHandle(args[Keys.ARG_DISPOSE_CALLBACK]));
+      disposeCallback();
     }
   });
   _backgroundChannel.invokeMethod(Keys.METHOD_SERVICE_INITIALIZED);